+++ scripts/os_scripts/linux.build.sh
@@ -54,7 +54,16 @@ cp -r ./dexcalibur ./dist/dexcalibur
 cp -r ./README.md ./dist/scripts
 cp -r ./src/requires ./dist/src/requires
 cp -r ./assets ./dist/assets
-cp -r ./thirdparty ./dist/thirdparty
+
+
+if [ ! -f ./dist/src/webserver/www ]; then
+  mkdir ./dist/src/webserver/www
+fi
+
+
+if [ -f ./thirdparty ]; then
+  cp -r ./thirdparty ./dist/thirdparty
+fi
 
 if [ -f ./public ]; then
   cp -r ./public ./dist/public
